+++ systemimage/service.py
@@ -95,7 +95,7 @@ def main():
         testing_mode = getattr(args, 'testing', None)
         if testing_mode:
             instrument(config, stack)
-            service = get_service(testing_mode, session_bus, '/Service', loop)
+            service = get_service(testing_mode, session_bus, '/Service')
             with open('/tmp/debug.log', 'a', encoding='utf-8') as fp:
                 print('MODE:', testing_mode, service, file=fp)
         else: