+++ systemimage/testing/dbus.py
@@ -16,7 +16,6 @@
 """Helpers for the DBus service when run with --testing."""
 
 __all__ = [
-    'TestableService',
     'get_service',
     'instrument',
     ]
@@ -27,7 +26,9 @@ import os
 from dbus.service import method
 from functools import partial
 from systemimage.api import Mediator
+from systemimage.config import config
 from systemimage.dbus import Service
+from systemimage.helpers import makedirs
 from systemimage.testing.helpers import test_data_path
 from unittest.mock import patch
 from urllib.request import urlopen
@@ -47,6 +48,7 @@ def instrument(config, stack):
     # Patch the subprocess call to write the reboot command to a log
     # file which the testing parent process can open and read.
     def safe_reboot(*args, **kws):
+        makedirs(config.updater.cache_partition)
         path = os.path.join(
             config.updater.cache_partition, 'reboot.log')
         with open(path, 'w', encoding='utf-8') as fp:
@@ -60,10 +62,6 @@ def instrument(config, stack):
 class _LiveTestableService(Service):
     """For testing purposes only."""
 
-    def __init__(self, bus, object_path, loop):
-        self._loop = loop
-        super().__init__(bus, object_path)
-
     @method('com.canonical.SystemImage')
     def Reset(self):
         self._api = Mediator()
@@ -72,9 +70,105 @@ class _LiveTestableService(Service):
         self._rebootable = True
 
 
-def get_service(testing_mode, session_bus, object_path, loop):
+class _NoUpdateService(Service):
+    """A static, 'no update is available' service."""
+
+    @method('com.canonical.SystemImage', out_signature='i')
+    def BuildNumber(self):
+        return 42
+
+    def _check_for_update(self):
+        self.UpdateAvailableStatus(False)
+
+    @method('com.canonical.SystemImage')
+    def Reset(self):
+        pass
+
+
+class _UpdateSuccessService(Service):
+    """A static, 'update available' service."""
+
+    def __init__(self, bus, object_path):
+        super().__init__(bus, object_path)
+        self._canceled = False
+
+    @method('com.canonical.SystemImage', out_signature='i')
+    def BuildNumber(self):
+        return 42
+    
+    def _check_for_update(self):
+        self.UpdateAvailableStatus(True)
+        return False
+
+    @method('com.canonical.SystemImage')
+    def Cancel(self):
+        self._canceled = True
+
+    @method('com.canonical.SystemImage', out_signature='x')
+    def GetUpdateSize(self):
+        return 1337 * 1024
+
+    @method('com.canonical.SystemImage', out_signature='i')
+    def GetUpdateVersion(self):
+        return 44
+
+    @method('com.canonical.SystemImage', out_signature='aa{ss}')
+    def GetDescriptions(self):
+        return [
+            {'description': 'Ubuntu Edge support',
+             'description-fr': "Support d'Ubuntu Edge",
+             'description-en': 'Initialise your Colour',
+             'description-en_US': 'Initialize your Color',
+            },
+            {'description': 'Flipped container with 200% faster boot'},
+            ]
+
+    def _complete_update(self):
+        if self._canceled:
+            self.Canceled()
+        else:
+            self.ReadyToReboot()
+        return False
+
+    @method('com.canonical.SystemImage')
+    def Reboot(self):
+        if self._canceled:
+            self.Canceled()
+        # The actual reboot is mocked to write a reboot log, so go ahead and
+        # call it.  The client will check the log.
+        config.hooks.reboot().reboot()
+
+    @method('com.canonical.SystemImage')
+    def Reset(self):
+        self._canceled = False
+        self._completing = False
+        self._checking = False
+
+
+class _UpdateFailedService(_UpdateSuccessService):
+    def _complete_update(self):
+        if self._canceled:
+            self.Canceled()
+        else:
+            self.UpdateFailed()
+        return False
+
+    @method('com.canonical.SystemImage')
+    def Reboot(self):
+        if self._canceled:
+            self.Canceled()
+        self.UpdateFailed()
+
+
+def get_service(testing_mode, session_bus, object_path):
     """Return the appropriate service class for the testing mode."""
     if testing_mode == 'live':
-        return _LiveTestableService(session_bus, object_path, loop)
+        return _LiveTestableService(session_bus, object_path)
+    elif testing_mode == 'no-update':
+        return _NoUpdateService(session_bus, object_path)
+    elif testing_mode == 'update-success':
+        return _UpdateSuccessService(session_bus, object_path)
+    elif testing_mode == 'update-failed':
+        return _UpdateFailedService(session_bus, object_path)
     else:
         raise RuntimeError('Invalid testing mode: {}'.format(testing_mode))